# Homework/openapi.yaml
openapi: 3.0.0
info:
  title: Game_Store
  description: Create, find invoices.
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
    description: Local dev server
paths:
...
  /console:
    get:
      summary: Retrieves details of all the consoles
      responses:
        '200':
          description: list of all the consoles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Console'
    put:
      summary: Update the console
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        '200':
          description: list of all the updated consoles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Console'
        '400':
          description: Bad Request
        '422':
          description: Missing a field
        '415':
          description: Unsupported media type, the request body should be json
    post:
      summary: Add the console
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Console'
      responses:
        '200':
          description: list of all the items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Console'
        '400':
          description: Bad Request
        '422':
          description: Missing a field
        '415':
          description: Unsupported media type, the request body should be json
  /console/{id}:
    get:
      summary: Retrieves details of the console with the specified id
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the console we are looking for
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: show the console with the specified id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Console'
        '400':
          description: Bad Request
        '404':
          description: Id not found
    delete:
      summary: Delete all the data related to the console with the specified id
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the console we are looking for
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: show the console with the specified id
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Bad Request
        '404':
          description: Id not found
  /console/{manufacturer}:
    get:
      summary: Retrieves details of the console with the specified manufacturer
      parameters:
        - name: manufacturer
          in: path
          required: true
          description: The manufaturer of the console we are looking for
          schema:
            type: string
      responses:
        '200':
          description: show the console with the specified manufacturer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Console'
        '400':
          description: Bad Request
        '404':
          description: Manufacture not found        
...
components:
  schemas:
...
    Console:
      type: array
      items:
        type: object
        properties:
          consoleId:
            type: integer
          model:
            type: string
          manufacturer:
            type: string
          memoryAmount:
            type: string
          processor:
            type: string
          price:
            type: number
          quantity:
            type: integer
        required:
          - model
          - manufacturer
          - memoryAmount
          - processor
          - price
          - quantity
...